RoundNumber->uint_64 ans some fixes

+++ lib/include/lib/system/common.hpp
@@ -16,7 +16,7 @@ constexpr std::size_t SIGNATURE_LENGTH = cscrypto::kSignatureSize;
 
 namespace cs {
 // key node type
-using RoundNumber = uint32_t;
+using RoundNumber = uint64_t;
 
 using Byte = cscrypto::Byte;
